fix permissions issue for new groups
